requirements.txt
# Azure Functions runtime (ASGI middleware wrapper)
azure-functions==1.21.3

# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# HTTP client
requests==2.32.5
httpx==0.28.1

# Azure authentication (DefaultAzureCredential, Managed Identity)
azure-identity==1.25.2

# Data validation / config models (Pydantic v2)
pydantic==2.12.5
pydantic-settings==2.9.1

# YAML parsing (CLI)
pyyaml==6.0.2

# JWT validation (Entra ID bearer tokens)
PyJWT[crypto]==2.10.1

# Structured JSON logging + coloured console output on Windows
structlog==24.2.0
colorama==0.4.6

# Circuit breaker
pybreaker==1.4.1

# Testing
pytest==8.3.5